rename value_slice.rs to values.rs
